requirements.txt
paho_mqtt
pycryptodome
aiohttp